Fix visualizer not working in other languages
Visualizers were not animating for some users. Old code would track if
the play button was on "Play" or "Paused" state. If the client was in
another language, these words would be different. New code is now
tracking a different node that seems to be consistent across all
client languages.