Update firefox mv2 manifest
This was set to "<all_urls>" because for the firefox version,
access was needed to the urls where the album images where stored
otherwise it would throw a CORS error and the palette picker wouldn't
work. Now I just included the youtube music url as well as the other
known album image hosting sites. There may be others.