Fix data merging issue on prefs update / extension update
When extension is updated with a new feature with its own prefs, the
extension would throw a catastrophic error which would need a full data
reset. It was taking the existing user prefs from the cloud store and
replacing the new prefs completely. Therefore the new preference for a
feature would be missing.
Now it's merging in the cloud data with the necessary data.